fix: isolate radar enhancement behind legacy renderer
