+++ skills/decision-first-dashboard/scripts/render.js
@@ -1,23 +1,7 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { validateDecisionState } from './validate.js';
+import { renderSvg as legacyRenderSvg, renderHtml as legacyRenderHtml } from './render-legacy.js';
 
-const currentFile = fileURLToPath(import.meta.url);
-const currentDir = path.dirname(currentFile);
-const noScoreSvgTemplatePath = path.resolve(currentDir, '../templates/no-score.svg');
-const noScoreHtmlTemplatePath = path.resolve(currentDir, '../templates/no-score.html');
-const compositeSvgTemplatePath = path.resolve(currentDir, '../templates/composite.svg');
-const compositeHtmlTemplatePath = path.resolve(currentDir, '../templates/composite.html');
-const cssTemplatePath = path.resolve(currentDir, '../templates/dashboard.css');
-
-function assertValid(data) {
-  const result = validateDecisionState(data);
-  if (!result.valid) {
-    const error = new Error('Decision state failed schema validation');
-    error.validationErrors = result.errors;
-    throw error;
-  }
+function formatScore(value) {
+  return Number.isInteger(value) ? String(value) : String(Number(value.toFixed(1)));
 }
 
 function escapeMarkup(value = '') {
@@ -29,477 +13,107 @@ function escapeMarkup(value = '') {
     .replaceAll("'", '&#39;');
 }
 
-export function deriveOverallDirection(signals = []) {
-  const known = new Set(
-    signals
-      .map((signal) => signal.direction)
-      .filter((direction) => direction && direction !== 'unknown')
-  );
-
-  if (known.size === 0) return 'unknown';
-  if (known.has('improving') && known.has('deteriorating')) return 'mixed';
-  if (known.has('improving')) return 'improving';
-  if (known.has('deteriorating')) return 'deteriorating';
-  if (known.has('flat')) return 'flat';
-  return 'unknown';
+function point(cx, cy, radius, angle) {
+  return { x: cx + Math.cos(angle) * radius, y: cy + Math.sin(angle) * radius };
 }
 
-function synthesisCopy(signals) {
-  const direction = {
-    improving: 'IMPROVING',
-    deteriorating: 'DETERIORATING',
-    mixed: 'MIXED',
-    flat: 'FLAT',
-    unknown: 'UNKNOWN'
-  }[deriveOverallDirection(signals)];
-
-  return { direction, target: 'Target unknown' };
+function pathFor(points) {
+  return points.map((p, i) => `${i === 0 ? 'M' : 'L'}${p.x.toFixed(1)} ${p.y.toFixed(1)}`).join(' ') + ' Z';
 }
 
-function directionClass(direction) {
-  if (direction === 'improving') return 'positive';
-  if (direction === 'deteriorating') return 'danger';
-  return 'muted';
-}
-
-function signalDisplay(signal) {
-  const hasMovement = Boolean(signal.delta);
-  return {
-    primary: hasMovement ? signal.delta : signal.value,
-    detail: hasMovement ? signal.value : null,
-    hasMovement
-  };
-}
-
-function selectRevenueSignal(signals) {
-  return signals.find((signal) => signal.metric === 'mrr')
-    ?? signals.find((signal) => signal.metric === 'arr')
-    ?? null;
-}
-
-function revenueTitle(signal) {
-  return signal ? `${signal.label} context` : 'Revenue context';
-}
-
-function currentRevenueLabel(signal) {
-  return signal ? `Current ${signal.label}` : 'Revenue metric unavailable';
-}
-
-function svgRevenueDeltaBlock(signal) {
-  if (!signal?.delta) return '';
-  return `<text x="98" y="252" class="${directionClass(signal.direction)}" font-size="14" font-weight="650">${escapeMarkup(signal.delta)}</text>`;
-}
-
-function htmlRevenueDeltaBlock(signal) {
-  if (!signal?.delta) return '';
-  return `<div class="metric-delta ${directionClass(signal.direction)}">${escapeMarkup(signal.delta)}</div>`;
-}
-
-function pathForSeries(series, { left, right, top, bottom }) {
-  const min = Math.min(...series);
-  const max = Math.max(...series);
+function geometry(data, { cx, cy, radius, labelRadius }) {
+  const { min, max } = data.score;
   const span = Math.max(max - min, 1);
+  const count = data.model.components.length;
+  const axes = [];
+  const inner = [];
+  const shape = [];
+  const labels = [];
 
-  return series
-    .map((value, index) => {
-      const x = left + ((right - left) * index) / (series.length - 1);
-      const y = bottom - ((value - min) / span) * (bottom - top);
-      return `${index === 0 ? 'M' : 'L'}${x.toFixed(1)} ${y.toFixed(1)}`;
-    })
-    .join(' ');
-}
-
-function svgRevenueVisual(series) {
-  if (!series?.length) {
-    return '<text x="98" y="330" class="muted" font-size="12">Trend data unavailable</text>';
-  }
-  const path = pathForSeries(series, { left: 98, right: 332, top: 286, bottom: 360 });
-  return `<path d="${path}" fill="none" stroke="url(#accentLine)" stroke-width="4" stroke-linecap="round" stroke-linejoin="round"/>`;
-}
-
-function htmlRevenueVisual(series) {
-  if (!series?.length) {
-    return '<div class="trend-unavailable">Trend data unavailable</div>';
-  }
-  const path = pathForSeries(series, { left: 0, right: 234, top: 6, bottom: 80 });
-  return `<svg class="sparkline" viewBox="0 0 234 86" aria-label="Revenue trend"><path d="${path}"></path></svg>`;
-}
-
-function svgScoreVisual(series) {
-  if (!series?.length) {
-    return '<text x="98" y="330" class="muted" font-size="12">Trend data unavailable</text>';
-  }
-  const path = pathForSeries(series, { left: 98, right: 332, top: 286, bottom: 360 });
-  return `<path d="${path}" fill="none" stroke="url(#accentLine)" stroke-width="4" stroke-linecap="round" stroke-linejoin="round"/>`;
-}
-
-function htmlScoreVisual(series) {
-  if (!series?.length) {
-    return '<div class="trend-unavailable">Trend data unavailable</div>';
-  }
-  const path = pathForSeries(series, { left: 0, right: 234, top: 6, bottom: 80 });
-  return `<svg class="sparkline" viewBox="0 0 234 86" aria-label="Score trend"><path d="${path}"></path></svg>`;
-}
-
-function splitSignalRows(signals) {
-  const topCount = Math.ceil(signals.length / 2);
-  return {
-    top: signals.slice(0, topCount),
-    bottom: signals.slice(topCount)
-  };
-}
-
-function svgRowXs(count) {
-  return {
-    1: [698],
-    2: [548, 848],
-    3: [478, 698, 918]
-  }[count] ?? [];
-}
-
-function htmlRowXs(count) {
-  return {
-    1: [50],
-    2: [26, 74],
-    3: [15, 50, 85]
-  }[count] ?? [];
-}
-
-function svgSignalCluster(signals) {
-  const { top, bottom } = splitSignalRows(signals);
-  const placed = [];
-
-  for (const [rowName, row, y] of [['top', top, 330], ['bottom', bottom, 590]]) {
-    const xs = svgRowXs(row.length);
-    row.forEach((signal, index) => placed.push({ signal, x: xs[index], y, rowName }));
-  }
-
-  const paths = placed
-    .map(({ x, y }) => `M698 464 L${x} ${y}`)
-    .join(' ');
-
-  const nodes = placed.map(({ signal, x, y, rowName }) => {
-    const topRow = rowName === 'top';
-    const primaryY = topRow ? y - 62 : y + 66;
-    const labelY = topRow ? y - 37 : y + 91;
-    const detailY = topRow ? y - 18 : y + 110;
-    const display = signalDisplay(signal);
-    const detailNode = display.detail
-      ? `<text x="${x}" y="${detailY}" class="muted" font-size="11" text-anchor="middle">${escapeMarkup(display.detail)}</text>`
-      : '';
-    return `<g class="signal-node">
-      <circle cx="${x}" cy="${y}" r="9" fill="#ffffff" stroke="#8d7fda" stroke-width="3"/>
-      <text x="${x}" y="${primaryY}" class="accent" font-size="24" font-weight="740" text-anchor="middle">${escapeMarkup(display.primary)}</text>
-      <text x="${x}" y="${labelY}" class="ink" font-size="13" font-weight="650" text-anchor="middle">${escapeMarkup(signal.label)}</text>
-      ${detailNode}
-    </g>`;
-  }).join('\n');
-
-  return { paths, nodes };
-}
-
-function htmlSignalCluster(signals) {
-  const { top, bottom } = splitSignalRows(signals);
-  const placed = [];
-
-  for (const [row, y] of [[top, 28], [bottom, 72]]) {
-    const xs = htmlRowXs(row.length);
-    row.forEach((signal, index) => placed.push({ signal, x: xs[index], y }));
-  }
-
-  const paths = placed.map(({ x, y }) => {
-    const px = (620 * x / 100).toFixed(1);
-    const py = (520 * y / 100).toFixed(1);
-    return `M310 260 L${px} ${py}`;
-  }).join(' ');
-
-  const nodes = placed.map(({ signal, x, y }) => {
-    const display = signalDisplay(signal);
-    const detailNode = display.detail ? `<small>${escapeMarkup(display.detail)}</small>` : '';
-    return `<div class="signal" style="left:${x}%;top:${y}%">
-    <strong>${escapeMarkup(display.primary)}</strong>
-    <span>${escapeMarkup(signal.label)}</span>
-    ${detailNode}
-  </div>`;
-  }).join('\n');
-
-  return { paths, nodes };
-}
-
-function formatScore(value) {
-  if (Number.isInteger(value)) return String(value);
-  return String(Number(value.toFixed(1)));
-}
-
-function formatWeight(weight) {
-  return `${Number((weight * 100).toFixed(1))}%`;
-}
-
-function radarPoint(cx, cy, radius, angle) {
-  return {
-    x: cx + Math.cos(angle) * radius,
-    y: cy + Math.sin(angle) * radius
-  };
-}
-
-function closedPath(points) {
-  return points
-    .map((point, index) => `${index === 0 ? 'M' : 'L'}${point.x.toFixed(1)} ${point.y.toFixed(1)}`)
-    .join(' ') + ' Z';
-}
-
-function radarGeometry(components, min, max, { cx, cy, radius, labelRadius }) {
-  const count = components.length;
-  const span = Math.max(max - min, 1);
-  const axisPoints = [];
-  const innerPoints = [];
-  const dataPoints = [];
-  const labelPoints = [];
-
-  components.forEach((component, index) => {
+  data.model.components.forEach((component, index) => {
     const angle = -Math.PI / 2 + (Math.PI * 2 * index) / count;
     const ratio = Math.min(1, Math.max(0, (component.normalizedScore - min) / span));
-    axisPoints.push(radarPoint(cx, cy, radius, angle));
-    innerPoints.push(radarPoint(cx, cy, radius * 0.5, angle));
-    dataPoints.push(radarPoint(cx, cy, radius * ratio, angle));
-    labelPoints.push(radarPoint(cx, cy, labelRadius, angle));
+    axes.push(point(cx, cy, radius, angle));
+    inner.push(point(cx, cy, radius * 0.5, angle));
+    shape.push(point(cx, cy, radius * ratio, angle));
+    labels.push(point(cx, cy, labelRadius, angle));
   });
 
   return {
-    grid: closedPath(axisPoints),
-    innerGrid: closedPath(innerPoints),
-    shape: closedPath(dataPoints),
-    spokes: axisPoints.map((point) => `M${cx} ${cy} L${point.x.toFixed(1)} ${point.y.toFixed(1)}`).join(' '),
-    axisPoints,
-    dataPoints,
-    labelPoints
+    grid: pathFor(axes),
+    inner: pathFor(inner),
+    shape: pathFor(shape),
+    spokes: axes.map((p) => `M${cx} ${cy} L${p.x.toFixed(1)} ${p.y.toFixed(1)}`).join(' '),
+    labels
   };
 }
 
-function svgComponentCluster(components, min, max) {
-  const geometry = radarGeometry(components, min, max, {
-    cx: 698,
-    cy: 464,
-    radius: 178,
-    labelRadius: 222
-  });
+function svgRadar(data) {
+  const g = geometry(data, { cx: 698, cy: 464, radius: 178, labelRadius: 222 });
+  return `<path class="radar-grid" d="${g.grid}" fill="rgba(114,100,202,0.015)" stroke="#d8d2ef" stroke-width="1.4"/>
+    <path class="radar-grid radar-grid--inner" d="${g.inner}" fill="none" stroke="#e9e5f6" stroke-width="1"/>
+    <path class="radar-spokes" d="${g.spokes}" fill="none" stroke="#dfdaf2" stroke-width="1.2"/>
+    <path class="radar-shape" d="${g.shape}" fill="rgba(114,100,202,0.20)" stroke="#7264ca" stroke-width="3" stroke-linejoin="round"/>`;
+}
 
-  const nodes = components.map((component, index) => {
-    const axis = geometry.axisPoints[index];
-    const label = geometry.labelPoints[index];
-    const detail = `${formatWeight(component.weight)} · ${component.value}`;
+function svgNodes(data) {
+  const g = geometry(data, { cx: 698, cy: 464, radius: 178, labelRadius: 222 });
+  return data.model.components.map((component, index) => {
+    const p = g.labels[index];
     return `<g class="score-component-node">
-      <circle cx="${axis.x.toFixed(1)}" cy="${axis.y.toFixed(1)}" r="8" fill="#ffffff" stroke="#8d7fda" stroke-width="3"/>
-      <text x="${label.x.toFixed(1)}" y="${(label.y - 12).toFixed(1)}" class="accent" font-size="22" font-weight="740" text-anchor="middle">${escapeMarkup(formatScore(component.normalizedScore))}</text>
-      <text x="${label.x.toFixed(1)}" y="${(label.y + 11).toFixed(1)}" class="ink" font-size="13" font-weight="650" text-anchor="middle">${escapeMarkup(component.label)}</text>
-      <text x="${label.x.toFixed(1)}" y="${(label.y + 30).toFixed(1)}" class="muted" font-size="11" text-anchor="middle">${escapeMarkup(detail)}</text>
-    </g>`;
-  }).join('\n');
-
-  return { ...geometry, nodes };
-}
-
-function htmlComponentCluster(components, min, max) {
-  const geometry = radarGeometry(components, min, max, {
-    cx: 310,
-    cy: 260,
-    radius: 174,
-    labelRadius: 216
-  });
-
-  const nodes = components.map((component, index) => {
-    const axis = geometry.axisPoints[index];
-    const label = geometry.labelPoints[index];
-    const left = (label.x / 620 * 100).toFixed(2);
-    const top = (label.y / 520 * 100).toFixed(2);
-    return `<div class="signal score-component" style="left:${left}%;top:${top}%">
-    <strong>${escapeMarkup(formatScore(component.normalizedScore))}</strong>
-    <span>${escapeMarkup(component.label)}</span>
-    <small>${escapeMarkup(`${formatWeight(component.weight)} · ${component.value}`)}</small>
-  </div>`;
-  }).join('\n');
-
-  return { ...geometry, nodes };
-}
-
-function svgCompositionRows(components) {
-  return components.map((component, index) => {
-    const y = 552 + index * 34;
-    const detail = `${formatScore(component.normalizedScore)} · ${formatWeight(component.weight)}`;
-    return `<g>
-      <text x="98" y="${y}" class="ink" font-size="12" font-weight="650">${escapeMarkup(component.label)}</text>
-      <text x="332" y="${y}" class="muted" font-size="11" text-anchor="end">${escapeMarkup(detail)}</text>
+      <text x="${p.x.toFixed(1)}" y="${(p.y - 12).toFixed(1)}" class="accent" font-size="22" font-weight="740" text-anchor="middle">${escapeMarkup(formatScore(component.normalizedScore))}</text>
+      <text x="${p.x.toFixed(1)}" y="${(p.y + 11).toFixed(1)}" class="ink" font-size="13" font-weight="650" text-anchor="middle">${escapeMarkup(component.label)}</text>
     </g>`;
   }).join('\n');
 }
 
-function htmlCompositionRows(components) {
-  return components.map((component) => `<div class="composition-row">
-    <span>${escapeMarkup(component.label)}</span>
-    <small>${escapeMarkup(`${formatScore(component.normalizedScore)} · ${formatWeight(component.weight)}`))}</small>
-  </div>`).join('\n');
+function enhanceSvg(markup, data) {
+  const spokePath = /<path d="M698 464[^\"]*" fill="none" stroke="#dad5ed" stroke-width="2"\/>/;
+  markup = markup.replace(spokePath, svgRadar(data));
+  markup = markup.replaceAll('r="98" fill="#ffffff"', 'r="76" fill="#ffffff"');
+  markup = markup.replaceAll('r="86" fill="#f7f4ff"', 'r="66" fill="#f7f4ff"');
+  markup = markup.replace(/\n\s*<g class="score-component-node">[\s\S]*?<\/g>/g, '');
+  const nodes = svgNodes(data);
+  return markup.replace(/\n\s*<g filter="url\(#cardShadow\)">\n\s*<rect x="1038"/, `\n    ${nodes}\n\n    <g filter="url(#cardShadow)">\n      <rect x="1038"`);
 }
 
-function movementSignals(signals, revenueMetric = null) {
-  const eligible = signals.filter((signal) => signal.delta);
-  const preferred = ['churn_rate', 'trial_conversion'];
-  const selected = [];
-
-  for (const metric of preferred) {
-    const match = eligible.find((signal) => signal.metric === metric);
-    if (match && !selected.includes(match)) selected.push(match);
-  }
-
-  for (const signal of eligible) {
-    if (selected.length >= 2) break;
-    if (signal.metric !== revenueMetric && !selected.includes(signal)) selected.push(signal);
-  }
-
-  return selected.slice(0, 2);
+function htmlRadar(data) {
+  const g = geometry(data, { cx: 310, cy: 260, radius: 174, labelRadius: 216 });
+  return `<svg viewBox="0 0 620 520" preserveAspectRatio="xMidYMid meet" aria-label="Component radar">
+      <path class="radar-grid" d="${g.grid}"></path>
+      <path class="radar-grid radar-grid--inner" d="${g.inner}"></path>
+      <path class="radar-spokes" d="${g.spokes}"></path>
+      <path class="radar-shape" d="${g.shape}"></path>
+    </svg>`;
 }
 
-function svgMovementRows(signals) {
-  if (signals.length === 0) {
-    return '<text x="98" y="566" class="muted" font-size="12">No additional movement signals</text>';
-  }
-
-  return signals.map((signal, index) => {
-    const labelY = 552 + index * 72;
-    const valueY = labelY + 27;
-    const divider = index === 0 && signals.length > 1
-      ? '<line x1="98" y1="600" x2="332" y2="600" stroke="#efedf6"/>'
-      : '';
-    return `<g>
-      <text x="98" y="${labelY}" class="muted" font-size="12">${escapeMarkup(signal.label)}</text>
-      <text x="98" y="${valueY}" class="ink" font-size="21" font-weight="700">${escapeMarkup(signal.value)}</text>
-      <text x="332" y="${valueY}" class="${directionClass(signal.direction)}" font-size="13" font-weight="650" text-anchor="end">${escapeMarkup(signal.delta)}</text>
-      ${divider}
-    </g>`;
-  }).join('\n');
-}
-
-function htmlMovementRows(signals) {
-  if (signals.length === 0) {
-    return '<div class="empty-state">No additional movement signals</div>';
-  }
-
-  return signals.map((signal) => `<div class="context-metric">
-    <span>${escapeMarkup(signal.label)}</span>
-    <strong>${escapeMarkup(signal.value)}</strong>
-    <em class="${directionClass(signal.direction)}">${escapeMarkup(signal.delta)}</em>
-  </div>`).join('\n');
-}
-
-function svgExceptionRows(exceptions = []) {
-  if (exceptions.length === 0) {
-    return '<text x="1064" y="232" class="muted" font-size="13">No confirmed exceptions visible</text>';
-  }
-
-  return exceptions.slice(0, 3).map((item, index) => {
-    const y = 232 + index * 70;
-    const meta = [item.plan, item.mrr ? `${item.mrr} MRR` : null].filter(Boolean).join(' · ');
-    return `<g>
-      <circle cx="1072" cy="${y - 4}" r="5" fill="#c96e79"/>
-      <text x="1088" y="${y}" class="ink" font-size="14" font-weight="650">${escapeMarkup(item.name)}</text>
-      <text x="1088" y="${y + 22}" class="muted" font-size="12">${escapeMarkup(meta)}</text>
-      <rect x="1268" y="${y - 19}" width="74" height="26" rx="13" fill="#fff1f3"/>
-      <text x="1305" y="${y - 1}" class="danger" font-size="11" font-weight="650" text-anchor="middle">${escapeMarkup(item.status)}</text>
-    </g>`;
-  }).join('\n');
-}
-
-function svgEventRows(events = []) {
-  if (events.length === 0) {
-    return '<text x="1064" y="534" class="muted" font-size="13">No recent source-supported events</text>';
-  }
-
-  return events.slice(0, 4).map((item, index) => {
-    const y = 534 + index * 68;
-    return `<g>
-      <circle cx="1072" cy="${y - 4}" r="4" fill="#8d7fda"/>
-      <text x="1088" y="${y}" class="ink" font-size="13" font-weight="650">${escapeMarkup(item.subject)}</text>
-      <text x="1088" y="${y + 21}" class="muted" font-size="12">${escapeMarkup(item.event)}</text>
-      <text x="1342" y="${y + 21}" class="soft" font-size="11" text-anchor="end">${escapeMarkup(item.time)}</text>
-    </g>`;
-  }).join('\n');
-}
-
-function htmlExceptionRows(exceptions = []) {
-  if (exceptions.length === 0) {
-    return '<div class="empty-state">No confirmed exceptions visible</div>';
-  }
-
-  return exceptions.slice(0, 3).map((item) => {
-    const meta = [item.plan, item.mrr ? `${item.mrr} MRR` : null].filter(Boolean).join(' · ');
-    return `<div class="exception-row">
-      <span class="dot dot--risk" aria-hidden="true"></span>
-      <div class="row-copy"><strong>${escapeMarkup(item.name)}</strong><span>${escapeMarkup(meta)}</span></div>
-      <span class="badge">${escapeMarkup(item.status)}</span>
+function htmlNodes(data) {
+  const g = geometry(data, { cx: 310, cy: 260, radius: 174, labelRadius: 216 });
+  return data.model.components.map((component, index) => {
+    const p = g.labels[index];
+    return `<div class="signal score-component" style="left:${(p.x / 620 * 100).toFixed(2)}%;top:${(p.y / 520 * 100).toFixed(2)}%">
+      <strong>${escapeMarkup(formatScore(component.normalizedScore))}</strong>
+      <span>${escapeMarkup(component.label)}</span>
     </div>`;
   }).join('\n');
 }
 
-function htmlEventRows(events = []) {
-  if (events.length === 0) {
-    return '<div class="empty-state">No recent source-supported events</div>';
-  }
+const RADAR_CSS = `
+.orbit path{vector-effect:non-scaling-stroke}.orbit .radar-grid{fill:rgba(114,100,202,.015);stroke:#d8d2ef;stroke-width:1.4}.orbit .radar-grid--inner{fill:none;stroke:#e9e5f6;stroke-width:1}.orbit .radar-spokes{fill:none;stroke:#dfdaf2;stroke-width:1.2}.orbit .radar-shape{fill:rgba(114,100,202,.22);stroke:var(--accent);stroke-width:3;stroke-linejoin:round}.score-center .synthesis-core{width:154px;height:154px}.score-component strong{font-size:24px}`;
 
-  return events.slice(0, 4).map((item) => `<div class="event-row">
-    <span class="dot" aria-hidden="true"></span>
-    <div class="row-copy"><strong>${escapeMarkup(item.subject)}</strong><span>${escapeMarkup(item.event)}</span></div>
-    <span class="event-time">${escapeMarkup(item.time)}</span>
-  </div>`).join('\n');
+function enhanceHtml(markup, data) {
+  markup = markup.replace(/<svg viewBox="0 0 620 520"[\s\S]*?<\/svg>/, htmlRadar(data));
+  markup = markup.replace(/\n\s*<div class="signal score-component"[\s\S]*?<\/div>/g, '');
+  markup = markup.replace('<div class="synthesis-core">', `${htmlNodes(data)}\n\n        <div class="synthesis-core">`);
+  return markup.replace('</style>', `${RADAR_CSS}\n</style>`);
 }
 
-function fillTemplate(template, replacements) {
-  let output = template;
-  for (const [key, value] of Object.entries(replacements)) {
-    output = output.replaceAll(`{{${key}}}`, value);
-  }
-  return output;
+export function renderSvg(data) {
+  const markup = legacyRenderSvg(data);
+  return data.mode === 'composite' ? enhanceSvg(markup, data) : markup;
 }
 
-function noScoreViewModel(data) {
-  const revenue = selectRevenueSignal(data.signals);
-  const synthesis = synthesisCopy(data.signals);
-  const revenueSeries = data.context?.provenance === 'source' ? data.context.revenueSeries : null;
-  const movement = movementSignals(data.signals, revenue?.metric ?? null);
-  return { revenue, synthesis, revenueSeries, movement };
+export function renderHtml(data) {
+  const markup = legacyRenderHtml(data);
+  return data.mode === 'composite' ? enhanceHtml(markup, data) : markup;
 }
-
-function renderNoScoreSvg(data) {
-  const template = fs.readFileSync(noScoreSvgTemplatePath, 'utf8');
-  const { revenue, synthesis, revenueSeries, movement } = noScoreViewModel(data);
-  const signalCluster = svgSignalCluster(data.signals);
-
-  return fillTemplate(template, {
-    REVENUE_TITLE: escapeMarkup(revenueTitle(revenue)),
-    REVENUE_VALUE: escapeMarkup(revenue?.value ?? '—'),
-    REVENUE_DELTA_BLOCK: svgRevenueDeltaBlock(revenue),
-    CURRENT_REVENUE_LABEL: escapeMarkup(currentRevenueLabel(revenue)),
-    REVENUE_VISUAL: svgRevenueVisual(revenueSeries),
-    MOVEMENT_ROWS: svgMovementRows(movement),
-    SYNTHESIS: synthesis.direction,
-    SVG_ORBIT_PATHS: signalCluster.paths,
-    SVG_SIGNAL_NODES: signalCluster.nodes,
-    EXCEPTION_ROWS: svgExceptionRows(data.exceptions),
-    EVENT_ROWS: svgEventRows(data.events)
-  });
-}
-
-function renderNoScoreHtml(data) {
-  const template = fs.readFileSync(noScoreHtmlTemplatePath, 'utf8');
-  const css = fs.readFileSync(cssTemplatePath, 'utf8');
-  const { revenue, synthesis, revenueSeries, movement } = noScoreViewModel(data);
-  const signalCluster = htmlSignalCluster(data.signals);
-
-  return fillTemplate(template, {
-    CSS: css,
-    REVENUE_TITLE: escapeMarkup(revenueTitle(revenue)),
-    REVENUE_VALUE: escapeMarkup(revenue?.value ?? '—'),
-    HTML_REVENUE_DELTA_BLOCK: htmlRevenueDeltaBlock(revenue),
-    CURRENT_REVENUE_LABEL: escapeMarkup(currentRevenueLabel(revenue)),
-    HTML_REVENUE_VISUAL: htmlRevenueVisual(revenueSeries),
-    HTML_MOVEMENT_ROWS: htmlMovementRows(movement),
-    SYNTHESIS: synthesis.direction,
-    HTML_ORBIT
